+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Updates Storybook and related dependencies to version 8.3.6.
 
+## Fixes
+
+- Fixes spacing issues and dark mode color styles in the `Table` component.
+
 ## 3.4.1 (October 24, 2024)
 
 ### Adds